+++ domain/src/main/java/ru/dpopkov/knownics/domain/KeyTerm.java
@@ -1,7 +1,12 @@
 package ru.dpopkov.knownics.domain;
 
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
 import ru.dpopkov.knownics.domain.user.AppUser;
 
+@Getter
+@Setter
 public class KeyTerm extends ModifiableEntity {
 
     private String name;
@@ -17,42 +22,11 @@ public class KeyTerm extends ModifiableEntity {
         this.description = description;
     }
 
+    @Builder
     public KeyTerm(String name, String description, AppUser createdBy, AppUser modifiedBy) {
         this.name = name;
         this.description = description;
         this.createdBy = createdBy;
-        this.modifiedBy = modifiedBy;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public AppUser getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(AppUser createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public AppUser getModifiedBy() {
-        return modifiedBy;
-    }
-
-    public void setModifiedBy(AppUser modifiedBy) {
         this.modifiedBy = modifiedBy;
     }