+++ domain/src/main/java/ru/dpopkov/knownics/domain/ModifiableComponent.java
@@ -1,25 +1,14 @@
 package ru.dpopkov.knownics.domain;
 
+import lombok.Getter;
+import lombok.Setter;
+
 import java.time.LocalDateTime;
 
+@Getter
+@Setter
 public abstract class ModifiableComponent {
 
     protected LocalDateTime created = LocalDateTime.now();
     protected LocalDateTime modified;
-
-    public LocalDateTime getCreated() {
-        return created;
-    }
-
-    public void setCreated(LocalDateTime created) {
-        this.created = created;
-    }
-
-    public LocalDateTime getModified() {
-        return modified;
-    }
-
-    public void setModified(LocalDateTime modified) {
-        this.modified = modified;
-    }
 }