+++ domain/src/main/java/ru/dpopkov/knownics/domain/question/Category.java
@@ -1,9 +1,13 @@
 package ru.dpopkov.knownics.domain.question;
 
+import lombok.Getter;
+import lombok.Setter;
 import ru.dpopkov.knownics.domain.ModifiableEntity;
 
 import javax.persistence.Entity;
 
+@Getter
+@Setter
 @Entity
 public class Category extends ModifiableEntity {
 
@@ -19,22 +23,6 @@ public class Category extends ModifiableEntity {
 
     public Category(String name, String description) {
         this.name = name;
-        this.description = description;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
         this.description = description;
     }