+++ domain/src/main/java/ru/dpopkov/knownics/domain/question/QCollection.java
@@ -1,11 +1,15 @@
 package ru.dpopkov.knownics.domain.question;
 
+import lombok.Getter;
+import lombok.Setter;
 import ru.dpopkov.knownics.domain.ModifiableEntity;
 import ru.dpopkov.knownics.domain.user.AppUser;
 
 import java.util.HashSet;
 import java.util.Set;
 
+@Getter
+@Setter
 public class QCollection extends ModifiableEntity {
 
     private String title;
@@ -13,46 +17,6 @@ public class QCollection extends ModifiableEntity {
     private AppUser owner;
     private boolean shared;
     private Set<Question> questions = new HashSet<>();
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public AppUser getOwner() {
-        return owner;
-    }
-
-    public void setOwner(AppUser owner) {
-        this.owner = owner;
-    }
-
-    public boolean isShared() {
-        return shared;
-    }
-
-    public void setShared(boolean shared) {
-        this.shared = shared;
-    }
-
-    public Set<Question> getQuestions() {
-        return questions;
-    }
-
-    public void setQuestions(Set<Question> questions) {
-        this.questions = questions;
-    }
 
     @Override
     public boolean equals(Object o) {