+++ history.md
@@ -3,6 +3,7 @@ History of implementation
 
 ##### 2021-12-29
 * Add i18n for web. Closes #14. 
+* Refactor to Lombok. Closes #15. 
 ##### 2021-12-28
 * Add module repository-jpa. Closes #11.
 * Add Profiles. Closes #12.